finish work on pdf
