+++ magnetopy/analyses/cauchy/plots.py
@@ -17,9 +17,35 @@ def plot_cauchy_pdf(
     x: npt.ArrayLike,
     y: npt.ArrayLike,
     results: CauchyAnalysisResults | None = None,
+    show_full_fit: bool = True,
+    show_fit_components: bool = False,
     input_params: CauchyFittingArgs | None = None,
     **kwargs,
 ) -> tuple[plt.Figure, plt.Axes]:
+    """Plots the results of a Cauchy PDF analysis.
+
+    Parameters
+    ----------
+    x : npt.ArrayLike
+        The x data, e.g., magnetic field.
+    y : npt.ArrayLike
+        They y data, e.g., derivative of magnetization with respect to magnetic field.
+    results : CauchyAnalysisResults | None, optional
+        The results of the fit, by default None.
+    show_full_fit : bool, optional
+        Assuming results is not None, whether to show the full fit, by default True.
+    show_fit_components : bool, optional
+        Assuming results is not None, whether to show the individual fit components,
+        by default False.
+    input_params : CauchyFittingArgs | None, optional
+        If given, plots the terms given by the CauchyFittingArgs object, by default
+        None.
+
+    Returns
+    -------
+    tuple[plt.Figure, plt.Axes]
+        The figure and axes objects.
+    """
     options = handle_kwargs(**kwargs)
 
     max_h_c = None
@@ -34,8 +60,13 @@ def plot_cauchy_pdf(
     if results:
         x_fit = np.linspace(min(x), max(x), 1000)
         y_fit = results.simulate_pdf_data(x_fit)
-        ax.plot(x_fit, y_fit, label="Fit", color="red")
         max_h_c = np.max([term.h_c for term in results.terms])
+        if show_full_fit:
+            ax.plot(x_fit, y_fit, label="Fit", color="red")
+        if show_fit_components:
+            y_fits = results.simulate_pdf_data_by_term(x_fit)
+            for i, y_fit in enumerate(y_fits):
+                ax.plot(x_fit, y_fit, label=f"Fit Term {i}")
 
     if options["xlabel"]:
         ax.set_xlabel(options["xlabel"])
@@ -55,7 +86,7 @@ def plot_cauchy_pdf(
         plt.savefig(
             options["save"], dpi=300, bbox_inches="tight", facecolor="w", edgecolor="w"
         )
-
+    ax.legend()
     return fig, ax
 
 
@@ -66,6 +97,8 @@ def _set_ax_lims(
     y: npt.ArrayLike,
     max_h_c: float | None,
 ) -> None:
+    x = np.array(x)
+    y = np.array(y)
     if max_h_c:
         if (np.max(x) - np.min(x)) > 50:
             # data is in Oe
